+++ src/citra_qt/bootmanager.h
@@ -104,7 +104,6 @@ private:
     std::unique_ptr<Input::ButtonDevice> capture_screenshot_then_send_to_discord_server_button;
     std::unique_ptr<Input::ButtonDevice> increase_volume_button;
     std::unique_ptr<Input::ButtonDevice> decrease_volume_button;
-    std::unique_ptr<Input::ButtonDevice> pause_button;
 
 signals:
     /**