+++ src/citra_qt/uisettings.h
@@ -102,7 +102,6 @@ struct Values {
     QString capture_screenshot_then_send_to_discord_server_button;
     QString increase_volume_button;
     QString decrease_volume_button;
-    QString pause_button;
 
     // Shortcut name <Shortcut, context>
     std::vector<Shortcut> shortcuts;